Добавлены логи. Добавлены новые эндпоинты. Добавлена логика выгрзуки в формате xml, xlsx, csv, json. Обновлены модели и методы.

+++ src/main/java/org/skeleton/team/repository/UplpDocRepository.java
@@ -3,5 +3,5 @@ package org.skeleton.team.repository;
 import org.skeleton.team.entity.UplpDoc;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UplpDocRepository extends JpaRepository<UplpDoc, Integer> {
+public interface UplpDocRepository extends JpaRepository<UplpDoc, Long> {
 }
